Add proper FCEUX label support
This isn't fully tested, but it should work fine.

Page files are generated dynamically and on-demand.  If there are no
labels in a page, that file will not be generated.

The FCEUX class handles all file output and generation (for FCEUX, not
Mesen) based on a given segment and symbol.