Save RAM needs to be handled differently
Save RAM is handled similarly to PRG ram in that it starts at $0000 in
the .mlb but with the "S" type instead of "P" or "R".  The value needs
to have $6000 subtracted from it to be the proper value when loaded
into Mesen.

TODO: test Save RAM labels in FCEUX